refactor IPC structure by consolidating workspace-related functions into a single module and removing deprecated API files
