--- view/desktop/src/tauriApi/describeLayoutPartsState.ts
@@ -1,6 +1,0 @@
-import { invokeTauriIpc } from "@/lib/backend/tauri";
-import { DescribeLayoutPartsStateOutput } from "@repo/moss-workspace";
-
-export const describeLayoutPartsState = async () => {
-  return await invokeTauriIpc<DescribeLayoutPartsStateOutput>("describe_layout_parts_state");
-};
