--- view/desktop/src/tauriApi/index.ts
@@ -1,3 +1,0 @@
-export * from "./describeLayoutPartsState";
-export * from "./openWorkspace";
-export * from "./setLayoutPartsState";
